#150 fixed type incoherence and added rest ivocations Controller

+++ bank/src/main/java/pt/ulisboa/tecnico/softeng/bank/services/local/dataobjects/BankOperationData.java
@@ -50,9 +50,7 @@ public class BankOperationData {
 		return this.value;
 	}
 
-	public void setValue(int value) {
-		this.value = value;
-	}
+	public void setValue(double value) { this.value = value; }
 
 	public DateTime getTime() {
 		return this.time;
